chorce: create getKline
