+++ src/bot/handlers/command.go
@@ -2,21 +2,12 @@ package handlers
 
 import (
 	// "context"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
-	"net/http"
-	"os"
 	"strconv"
 	"strings"
 	"telegram-bot/services"
 
-	"time"
-
-	// "github.com/chromedp/chromedp"
-	"github.com/go-echarts/go-echarts/v2/charts"
-	"github.com/go-echarts/snapshot-chromedp/render"
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 )
 
@@ -308,93 +299,4 @@ func sendMenu(chatID int64) tgbotapi.MessageConfig {
 func copyMessage(message *tgbotapi.Message) tgbotapi.MessageConfig {
 	msg := tgbotapi.NewMessage(message.Chat.ID, message.Text)
 	return msg
-}
-
-func getKlineData(symbol string, interval string, options ...int) ([]klineData, error) {
-	apiURL := fmt.Sprintf("https://api.binance.com/api/v3/klines?symbol=%s&interval=%s", symbol, interval)
-
-	if len(options) > 0 {
-		apiURL = fmt.Sprintf("%s&limit=%d", apiURL, options[0])
-	}
-
-	resp, err := http.Get(apiURL)
-	if err != nil {
-		log.Printf("failed to make request: %v", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		log.Printf("request failed: %s", resp.Status)
-	}
-
-	var data [][]interface{}
-	body, _ := ioutil.ReadAll(resp.Body)
-	json.Unmarshal(body, &data)
-
-	var kd []klineData
-	for _, k := range data {
-		openTime := int64(k[0].(float64)) / 1000
-		date := time.Unix(openTime, 0).Format("2006-01-02 15:04:05")
-
-		open, _ := parseFloat32(k[1].(string))
-		high, _ := parseFloat32(k[2].(string))
-		low, _ := parseFloat32(k[3].(string))
-		close, _ := parseFloat32(k[4].(string))
-
-		kd = append(kd, klineData{
-			Date: date,
-			Data: [4]float32{open, close, low, high},
-		})
-	}
-
-	return kd, nil
-}
-
-func sendChartToTelegram(bot *tgbotapi.BotAPI, chatID int64, chart *charts.Kline) error {
-	initialMsg := tgbotapi.NewMessage(chatID, "Uploading file...")
-	sentMsg, err := bot.Send(initialMsg)
-	if err != nil {
-		log.Printf("failed to send initial message: %v", err)
-		return fmt.Errorf("failed to send initial message: %w", err)
-	}
-	log.Println("Uploading file...")
-
-	fileName := fmt.Sprintf("chart-%d%d.png", chatID, time.Now().UnixNano())
-	log.Printf("File name generated: %s", fileName)
-
-	// Render chart with Chromedp context
-	err = render.MakeChartSnapshot(chart.RenderContent(), fileName)
-	if err != nil {
-		log.Printf("Failed to generate chart snapshot: %v", err)
-		return fmt.Errorf("failed to generate chart snapshot: %w", err)
-	}
-	log.Printf("Chart snapshot generated: %s", fileName)
-
-	imgBytes, err := ioutil.ReadFile(fileName)
-	if err != nil {
-		log.Printf("failed to read generated chart image: %v", err)
-		return fmt.Errorf("failed to read generated chart image: %w", err)
-	}
-	log.Printf("Image read successfully, size: %d bytes", len(imgBytes))
-
-	msg := tgbotapi.NewPhoto(chatID, tgbotapi.FileBytes{
-		Name:  fileName,
-		Bytes: imgBytes,
-	})
-
-	if _, err := bot.Send(msg); err != nil {
-		log.Printf("failed to send chart image: %v", err)
-		return fmt.Errorf("failed to send chart image: %w", err)
-	}
-
-	if err := os.Remove(fileName); err != nil {
-		log.Printf("warning: failed to delete image file %s: %v", fileName, err)
-	}
-
-	editMsg := tgbotapi.NewEditMessageText(chatID, sentMsg.MessageID, "File uploaded successfully!")
-	if _, err := bot.Send(editMsg); err != nil {
-		log.Printf("warning: failed to edit initial message: %v", err)
-	}
-
-	return nil
 }
